==> 1.data_pre_processing.rmd <==
---
title: "Predictors of Depression During the Covid-19 Pandemic" 
subtitle: "Data Pre-processing"
author: "Sarka Tesarova, Ondrej Pekacek, Alessandro Porrovecchio"
date: "Last edited `r format (Sys.time(),'%d. %m. %Y')`"
output:
  html_document: 
    toc: yes
    toc_depth: 2
    toc_float: true
    number_sections: true
    theme: readable
    code_folding: hide
    code_download: true
    includes:
      in_header: docs/header.html
---

```{r loading packages and dataset, message=FALSE, warning=FALSE}
# The following packages might need to be installed onto your version 
# of R prior to the running of the code below.

# Package names
packages <- c("dplyr", "haven", "mice", "tidyr", "fastDummies")

# Install packages not yet installed
installed_packages <- packages %in% rownames(installed.packages())
if (any(installed_packages == FALSE)) {
  install.packages(packages[!installed_packages])
}

# Packages loading
invisible(lapply(packages, library, character.only = TRUE))

# We load the international dataset (in SPSS format) from a local directory.
# We also remove labels
data <- zap_labels(haven::read_sav(file = "data/international_dataset.sav"))

# We also try to limit the decimals to three significant figures
options(digits = 3, scipen = 999)

```

## Calculating PHQ8 dependent variable

```{r}
# To get the comparability with the original scale (0-24 points), we need to deduce 1
data$PHQ8 <- (data$Q49_1 - 1) + (data$Q49_2 - 1) + (data$Q49_3 - 1) + (data$Q49_4 - 1) + (data$Q49_5 - 1) + (data$Q49_6 - 1) + (data$Q49_7 - 1) + (data$Q49_8 - 1)

```

## Seting factor levels to be interpretable in the analysis, preserving only variables of interest

```{r}
data_subset <- data %>% 
  transmute(country = recode_factor(as_factor(Country),
            `1` = "IT",
            `2` = "FR",
            `3` = "PL",
            `4` = "UK",
            `5` = "CZ",
            `6` = "SW", .ordered = FALSE),
            q01_gender = recode_factor(as_factor(Q3),
            `1` = "male",
            `2` = "female",
            `3` = "other",
            .ordered = FALSE),
            q02_age = Q4,
            q02_age_group = recode_factor(as_factor(Q4_AGE_r),
            `1` = "16-29 years",
            `2` = "30-49 years",
            `3` = "50-64 years",
            `4` = "65+",
            .ordered = TRUE),
            q03_relationship_type = recode_factor(as_factor(Q5),
            `1` = "single",
            `2` = "relationship",
            `3` = "married",
            `4` = "divorced",
            `5` = "widowed",
            .ordered = FALSE),
            q04_children = recode_factor(as_factor(Q6),
            `1` = "yes",
            `2` = "no",
            .ordered = FALSE),
            q11_education = recode_factor(as_factor(Q13R),
            `1` = "low",
            `2` = "medium",
            `3` = "high",
            .ordered = TRUE),
            q18_02_soc_media = recode_factor(as_factor(replace_na(Q21_2, 0)),
            `0` = "no",
            `1` = "yes"
            ,.ordered = FALSE),
            q20_public_info = recode_factor(as_factor(Q23),
            `1` = "yes",
            `2` = "no",
            `99` = "do_not_know",
            .ordered = FALSE),
            q34_02_face_mask = recode_factor(as_factor(Q37_2),
            `1` = "yes",
            `2` = "no",
            .ordered = FALSE),
            q34_07_hand_washing = recode_factor(as_factor(Q37_7),
            `1` = "yes",
            `2` = "no",
            .ordered = FALSE),
            q35_01_contact_close_family = recode_factor(as_factor(Q38_1),
            `1` = "less_often",
            `2` = "as_before",
            `3` = "more_often",
            .ordered = TRUE),
            q35_03_contact_friends = recode_factor(as_factor(Q38_3),
            `1` = "less_often",
            `2` = "as_before",
            `3` = "more_often",
            .ordered = TRUE),
            q36_econ_worry = recode_factor(as_factor(Q39), 
            `1` = "very_serious",
            `2` = "serious",
            `3` = "limited",
            .ordered = TRUE),
            q38_alcohol = recode_factor(as_factor(Q42),
            `1` = "yes",
            `2` = "no",
            .ordered = FALSE),
            q40_smoking = recode_factor(as_factor(Q44),
            `1` = "yes",
            `2` = "no",
            .ordered = FALSE),
            q42_sport = recode_factor(as_factor(Q46),
            `1` = "yes",
            `2` = "no",
            .ordered = FALSE),
            q47_self_reporting_health = recode_factor(as_factor(Q50),
            `1` = "excellent",
            `2` = "good",
            `3` = "neutral",
            `4` = "bad",
            `5` = "very_bad",
            .ordered = TRUE),
            q48_chronic_illness = recode_factor(as_factor(Q51),
            `1` = "yes",
            `2` = "no",
            .ordered = FALSE),
            q49_health_limitations = recode_factor(as_factor(Q52),
            `1` = "limits",
            `2` = "partially_limits",
            `3` = "no_limits",
            .ordered = TRUE),
            q30_concern_infection_covid = Q33,
            q31_concern_infection_friends = Q34,
            q33_01_concern_situation = Q36_1,
            q33_02_concern_low_control = Q36_2,
            q33_03_concern_survival_covid = Q36_3,
            q33_04_concern_change_employment = Q36_4,
            q33_05_concern_infecting_others = Q36_5,
            PHQ8 = PHQ8) 

```

## Filtering out the Czech sample for analyses
```{r}
# Change Do Not Know to NA's
levels(data_subset$q20_public_info)[levels(data_subset$q20_public_info) == "do_not_know"] <- NA

# Save the current dataset version, which still contains country and age group
# This is useful for exploratory data analysis
saveRDS(data_subset, "data/eda_dataset.rds")

data_subset <- data_subset %>%
        filter(!country %in% c("CZ")) %>% 
        select(-c("country", "q02_age_group"))

```

## Analysis of the missing data
```{r}
# Dataset overview
summary(data_subset)

# What proportion are the missing data? (in percent)
(sum(is.na(data_subset)) / sum(!is.na(data_subset))) * 100

# What columns are most affected by missingness (in percent)
sort(colSums(is.na(data_subset)) / colSums(!is.na(data_subset)) * 100)

# Visualize patterns of missingness
md.pattern(data_subset)
```

## Multiple imputation using the MICE package

How many imputations? Rule of thumb (https://onlinelibrary.wiley.com/doi/10.1002/sim.4067), at least the number of missing in percent.
Recommendation of 10 iterations per imputation (https://academic.oup.com/aje/article/169/9/1133/125871)
```{r}
imputed_df <- mice(data_subset,
  m = 5,
  maxit = 10,
  defaultMethod = c("pmm", "logreg", "polyreg", "polr"),
  seed = 4167,
  print = FALSE
)
```

```{r}
# Remove factor ordering for better interpretation within regression coefficients, setting contrasts
imputed_df_reg <- complete(imputed_df, action = "long", include = TRUE) %>% 
   mutate(across(where(is.ordered), ~ factor(., ordered = FALSE)),
          q01_gender = relevel(q01_gender, ref = "male"),
          q03_relationship_type = relevel(q03_relationship_type, ref = "married"),
          q04_children = relevel(q04_children, ref = "yes"),
          q11_education = relevel(q11_education, ref = "high"),
          q18_02_soc_media = relevel(q18_02_soc_media, ref = "no"),
          q20_public_info = relevel(q20_public_info, ref = "yes"),
          q34_02_face_mask = relevel(q34_02_face_mask, ref = "yes"),
          q34_07_hand_washing = relevel(q34_07_hand_washing, ref = "yes"),
          q35_01_contact_close_family = relevel(q35_01_contact_close_family, ref = "more_often"),
          q35_03_contact_friends = relevel(q35_03_contact_friends, ref = "more_often"),
          q36_econ_worry = relevel(q36_econ_worry, ref = "limited"),
          q38_alcohol = relevel(q38_alcohol, ref = "no"),
          q40_smoking = relevel(q40_smoking, ref = "no"),
          q42_sport = relevel(q42_sport, ref = "yes"),
          q47_self_reporting_health = relevel(q47_self_reporting_health, ref = "excellent"),
          q48_chronic_illness = relevel(q48_chronic_illness, ref = "no"),
          q49_health_limitations = relevel(q49_health_limitations, ref = "no_limits")) %>%
   as.mids()

imputed_df_cor <- complete(imputed_df, action = "long", include = TRUE) %>% 
   dummy_cols(select_columns = c("q03_relationship_type", "q01_gender"), remove_selected_columns = TRUE) %>% 
   mutate(across(everything(), ~ as.numeric(.))) %>% 
   select(-c("q03_relationship_type_NA", "q01_gender_NA")) %>% 
   as.mids()

imputed_df_clust <- complete(imputed_df, action = "long", include = TRUE) %>% 
   mutate(q01_gender_female = factor(if_else(q01_gender == "female", 1, 0), levels = c(1, 0), labels = c("yes", "no")),
          q03_relationship_or_married = factor(if_else(q03_relationship_type %in% c("married", "relationship"), 1, 0), levels = c(1, 0), labels = c("yes", "no"))) %>% 
   transmute(.imp,
             .id,
             PHQ8,
             q01_gender_female,
             q02_age,
             q03_relationship_or_married,
             q04_children,
             q11_education,
             q18_02_soc_media,
             q20_public_info,
             q36_econ_worry,	
             q40_smoking,
             q42_sport,
             q47_self_reporting_health,
             q49_health_limitations) %>% 
  as.mids()

# We reverse code some of the categorical variables for more intuitive interpretation in the cluster analysis

imputed_df_clust_no_fct <- complete(imputed_df_clust, action = "long", include = TRUE) %>% 
      mutate(q01_gender_female = as.numeric(as.character(recode(q01_gender_female,
             "yes" = "1",
             "no" = "0"))),
             q03_relationship_or_married = as.numeric(as.character(recode(q03_relationship_or_married,
             "yes" = "1",
             "no" = "0"))),
             q04_children = as.numeric(as.character(recode(q04_children,
             "yes" = "1",
             "no" = "0"))),             
             q11_education = as.numeric(as.character(recode(q11_education,
             "low" = "1",
             "medium" = "2",
             "high" = "3"))),                  
             q18_02_soc_media = as.numeric(as.character(recode(q18_02_soc_media,
             "yes" = "1",
             "no" = "0"))),  
             q20_public_info = as.numeric(as.character(recode(q20_public_info,
             "yes" = "1",
             "no" = "0"))), 
             q36_econ_worry = as.numeric(as.character(recode(q36_econ_worry,
             "very_serious" = "3",
             "serious" = "2",
             "limited" = "1"))),
             q40_smoking = as.numeric(as.character(recode(q40_smoking,
             "yes" = "1",
             "no" = "0"))),
             q42_sport = as.numeric(as.character(recode(q42_sport,
             "yes" = "1",
             "no" = "0"))),
             q47_self_reporting_health = as.numeric(as.character(recode(q47_self_reporting_health,
             "excellent" = "5",
             "good" = "4",
             "neutral" = "3",
             "bad" = "2",
             "very_bad" = "1"))),
             q49_health_limitations = as.numeric(as.character(recode(q49_health_limitations,
             "limits" = "3",
             "partially_limits" = "2",
             "no_limits" = "1")))) %>% 
  as.mids()

```


## Saving the imputed dataset
```{r}
saveRDS(imputed_df, "data/imputed_dataset.rds")
saveRDS(imputed_df_reg, "data/imputed_dataset_reg.rds")
saveRDS(imputed_df_cor, "data/imputed_dataset_corr.rds")
saveRDS(imputed_df_clust, "data/imputed_dataset_clust.rds")
saveRDS(imputed_df_clust_no_fct, "data/imputed_dataset_clust_no_fct.rds")


```
