==> 2.multiple_regression.rmd <==
---
title: "Predictors of Depression During the Covid-19 Pandemic" 
subtitle: "Multiple Regression"
author: "Sarka Tesarova, Ondrej Pekacek, Alessandro Porrovecchio"
date: "Last edited `r format (Sys.time(),'%d. %m. %Y')`"
output:
  html_document: 
    toc: yes
    toc_depth: 2
    toc_float: true
    number_sections: true
    theme: readable
    code_folding: hide
    code_download: true
    includes:
      in_header: docs/header.html
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
```

```{r loading packages and dataset, message=FALSE, include=FALSE}
# The following packages might need to be installed onto your version 
# of R prior to the running of the code below.

# Package names
packages <- c("corrplot", "dplyr", "mice", "car", "miceadds", "kableExtra")

# Install packages not yet installed
installed_packages <- packages %in% rownames(installed.packages())
if (any(installed_packages == FALSE)) {
  install.packages(packages[!installed_packages])
}

# Packages loading
invisible(lapply(packages, library, character.only = TRUE))

imputed_df <- readRDS("data/imputed_dataset.rds")
imputed_df_corr <- readRDS("data/imputed_dataset_corr.rds")

# We also try to limit the decimals to three significant figures
options(digits = 3, scipen = 999)

```

# Correlation plot (Spearman, multiply imputed dataset)

```{r echo=FALSE, fig.height=9, fig.width=9}
mi_corr <- micombine.cor(imputed_df_corr, 
              method = "spearman", 
              conf.level = 0.95)

# Add significance matrix
corrplot(attr(mi_corr,"r_matrix"), 
         method = "circle", 
             type = "lower", 
             # p.mat = res1$p, 
             # sig.level = .05, 
             mar = c(0,0,1,0))

```


# Regression models (multiply imputed dataset) {.tabset .tabset-pills}

```{r include=FALSE}
# outcome <- "PHQ8"
# predictors <- c("q01_gender", "q02_age", "q04_children", "q36_econ_worry", "q18_02_soc_media", "q47_self_reporting_health", "q49_health_limitations")
# 
# formula_bic <- as.formula(
#   paste(outcome,
#     paste(predictors, collapse = " + "),
#     sep = " ~ "
#   )
# )

# test_lm <- lm(formula_bic, mice::complete(imputed_df))
# test <- imputed_df %>% 
#    mice::complete("all") %>%
#   lapply(lm, formula = PHQ8 ~ q01_gender + q02_age + q04_children + q36_econ_worry + 
#     q18_02_soc_media + q47_self_reporting_health + q49_health_limitations) %>%
#   lapply(Anova) 
# car::durbinWatsonTest(stepwise_BIC_fit)
# car::vif(test_lm)
# car::Anova(test_lm, p.adjust.method = TRUE)
# 

# test_jmv <- jmv::linReg(
#     data = data_subset,
#     dep = "PHQ8",
#     covs = "q02_age",
#     factors = vars("q01_gender"),
#      blocks = list(
#         list(
#             "q01_gender",
#             "q02_age")),
#     refLevels = list(
#         list(
#             var = "q01_gender",
#             ref = "female")),
#     r2Adj = TRUE,
#     aic = TRUE,
#     bic = TRUE,
#     rmse = TRUE,
#     modelTest = TRUE,
#     anova = TRUE,
#     ci = TRUE,
#     stdEst = TRUE,
#     ciStdEst = TRUE,
#     durbin = TRUE,
#     collin = TRUE)

# Testing model chosen based by stepwise algorithm on its BIC score
stepwise_BIC_fit <- with(imputed_df, lm(PHQ8 ~ q01_gender +
  q02_age +
  q04_children +
  q36_econ_worry +
  q18_02_soc_media +
  q47_self_reporting_health +
  q49_health_limitations)) %>%
  pool()

stepwise_BIC_anova <- mi.anova(imputed_df, "PHQ8 ~ q01_gender +
  q02_age +
  q04_children +
  q36_econ_worry +
  q18_02_soc_media +
  q47_self_reporting_health +
  q49_health_limitations")

# Testing model chosen based by stepwise algoritm on its AIC score
stepwise_AIC_fit <- with(imputed_df, lm(PHQ8 ~ q01_gender +
  q02_age +
  q04_children +
  q36_econ_worry +
  q18_02_soc_media +
  q47_self_reporting_health +
  q49_health_limitations +
  q03_relationship_type +
  q20_public_info +
  q34_02_face_mask +
  q38_alcohol +
  q40_smoking +
  q48_chronic_illness)) %>%
  pool()

stepwise_AIC_anova <- mi.anova(imputed_df, "PHQ8 ~ q01_gender +
  q02_age +
  q04_children +
  q36_econ_worry +
  q18_02_soc_media +
  q47_self_reporting_health +
  q49_health_limitations +
  q03_relationship_type +
  q20_public_info +
  q34_02_face_mask +
  q38_alcohol +
  q40_smoking +
  q48_chronic_illness")


BIC_summ <- summary(stepwise_BIC_fit, conf.int = TRUE) %>%
  mutate(p.adjusted = round(p.adjust(p.value, method = "holm"), 3)) %>%
  select(-p.value) %>% 
  arrange(desc(abs(estimate)))

AIC_summ <- summary(stepwise_AIC_fit, conf.int = TRUE) %>%
  mutate(p.adjusted = round(p.adjust(p.value, method = "holm"), 3)) %>%
  select(-p.value) %>% 
  arrange(desc(abs(estimate)))

BIC_summ_anv <- stepwise_BIC_anova$anova.table %>% arrange(desc(`F value`))
AIC_summ_anv <- stepwise_AIC_anova$anova.table %>% arrange(desc(`F value`))


```

```{r include=FALSE}
th_1_fit <- with(imputed_df, lm(PHQ8 ~ q01_gender +
  q02_age +
  q03_relationship_type +  
  q04_children +
  q11_education)) %>%
  pool()

th_2_fit <- with(imputed_df, lm(PHQ8 ~ q01_gender +
  q02_age +
  q03_relationship_type +  
  q04_children +
  q11_education +
  q18_02_soc_media +
  q20_public_info +
  q34_02_face_mask +
  q34_07_hand_washing +
  q36_econ_worry)) %>%
  pool()

th_3_fit <- with(imputed_df, lm(PHQ8 ~ q01_gender +
  q02_age +
  q03_relationship_type +  
  q04_children +
  q11_education +
  q18_02_soc_media +
  q20_public_info +
  q34_02_face_mask +
  q34_07_hand_washing +
  q36_econ_worry +
  q40_smoking +
  q42_sport +
  q38_alcohol +
  q35_01_contact_close_family +
  q35_03_contact_friends)) %>%
  pool()

th_4_fit <- with(imputed_df, lm(PHQ8 ~ q01_gender +
  q02_age +
  q03_relationship_type +  
  q04_children +
  q11_education +
  q18_02_soc_media +
  q20_public_info +
  q34_02_face_mask +
  q34_07_hand_washing +
  q36_econ_worry +
  q40_smoking +
  q42_sport +
  q38_alcohol +
  q35_01_contact_close_family +
  q35_03_contact_friends + 
  q47_self_reporting_health +
  q48_chronic_illness +
  q49_health_limitations)) %>%
  pool()

th_1_anv <- mi.anova(imputed_df, "PHQ8 ~ q01_gender +
  q02_age +
  q03_relationship_type +  
  q04_children +
  q11_education")

th_2_anv <- mi.anova(imputed_df, "PHQ8 ~ q01_gender +
  q02_age +
  q03_relationship_type +  
  q04_children +
  q11_education +
  q18_02_soc_media +
  q20_public_info +
  q34_02_face_mask +
  q34_07_hand_washing +
  q36_econ_worry")

th_3_anv <- mi.anova(imputed_df, "PHQ8 ~ q01_gender +
  q02_age +
  q03_relationship_type +  
  q04_children +
  q11_education +
  q18_02_soc_media +
  q20_public_info +
  q34_02_face_mask +
  q34_07_hand_washing +
  q36_econ_worry +
  q40_smoking +
  q42_sport +
  q38_alcohol +
  q35_01_contact_close_family +
  q35_03_contact_friends")

th_4_anv <- mi.anova(imputed_df, "PHQ8 ~ q01_gender +
  q02_age +
  q03_relationship_type +  
  q04_children +
  q11_education +
  q18_02_soc_media +
  q20_public_info +
  q34_02_face_mask +
  q34_07_hand_washing +
  q36_econ_worry +
  q40_smoking +
  q42_sport +
  q38_alcohol +
  q35_01_contact_close_family +
  q35_03_contact_friends + 
  q47_self_reporting_health +
  q48_chronic_illness +
  q49_health_limitations")

th_1_summ <- summary(th_1_fit, conf.int = TRUE) %>%
  mutate(p.adjusted = round(p.adjust(p.value, method = "holm"), 3)) %>%
  select(-p.value) %>% 
  arrange(desc(abs(estimate)))

th_2_summ <- summary(th_2_fit, conf.int = TRUE) %>%
  mutate(p.adjusted = round(p.adjust(p.value, method = "holm"), 3)) %>%
  select(-p.value) %>% 
  arrange(desc(abs(estimate)))

th_3_summ <- summary(th_3_fit, conf.int = TRUE) %>%
  mutate(p.adjusted = round(p.adjust(p.value, method = "holm"), 3)) %>%
  select(-p.value) %>% 
  arrange(desc(abs(estimate)))

th_4_summ <- summary(th_4_fit, conf.int = TRUE) %>%
  mutate(p.adjusted = round(p.adjust(p.value, method = "holm"), 3)) %>%
  select(-p.value) %>% 
  arrange(desc(abs(estimate)))

th_1_summ_anv <- th_1_anv$anova.table %>% arrange(desc(`F value`))
th_2_summ_anv <- th_2_anv$anova.table %>% arrange(desc(`F value`))
th_3_summ_anv <- th_3_anv$anova.table %>% arrange(desc(`F value`))
th_4_summ_anv <- th_4_anv$anova.table %>% arrange(desc(`F value`))

```

## Model fit measures

```{r} 
kbl(pool.r.squared(stepwise_BIC_fit), caption = "Stepwise BIC Model: Fit Measures") %>%
kable_classic(full_width = FALSE, lightable_options = c("striped"))

kbl(pool.r.squared(stepwise_AIC_fit), caption = "Stepwise AIC Model: Fit Measures") %>%
kable_classic(full_width = FALSE, lightable_options = c("striped"))

kbl(pool.r.squared(th_1_fit), caption = "Theory-derived Model 1: Fit Measures") %>%
kable_classic(full_width = FALSE, lightable_options = c("striped"))

kbl(pool.r.squared(th_2_fit), caption = "Theory-derived Model 2: Fit Measures") %>%
kable_classic(full_width = FALSE, lightable_options = c("striped"))

kbl(pool.r.squared(th_3_fit), caption = "Theory-derived Model 3: Fit Measures") %>%
kable_classic(full_width = FALSE, lightable_options = c("striped"))

kbl(pool.r.squared(th_4_fit), caption = "Theory-derived Model 4: Fit Measures") %>%
kable_classic(full_width = FALSE, lightable_options = c("striped"))

```

## Model coefficients

```{r}
kbl(BIC_summ, caption = "Stepwise BIC Model: Coefficients") %>%
kable_classic(full_width = FALSE, lightable_options = c("striped")) %>%
                    row_spec(row = which(BIC_summ$p.adjusted < 0.05), bold = TRUE)

kbl(AIC_summ, caption = "Stepwise AIC Model: Coefficients") %>%
kable_classic(full_width = FALSE, lightable_options = c("striped")) %>%
                    row_spec(row = which(AIC_summ$p.adjusted < 0.05), bold = TRUE)

kbl(th_1_summ, caption = "Theory-derived Model 1: Coefficients") %>%
kable_classic(full_width = FALSE, lightable_options = c("striped")) %>%
                    row_spec(row = which(th_1_summ$p.adjusted < 0.05), bold = TRUE)

kbl(th_2_summ, caption = "Theory-derived Model 2: Coefficients") %>%
kable_classic(full_width = FALSE, lightable_options = c("striped")) %>%
                    row_spec(row = which(th_2_summ$p.adjusted < 0.05), bold = TRUE)

kbl(th_3_summ, caption = "Theory-derived Model 3: Coefficients") %>%
kable_classic(full_width = FALSE, lightable_options = c("striped")) %>%
                    row_spec(row = which(th_3_summ$p.adjusted < 0.05), bold = TRUE)

kbl(th_4_summ, caption = "Theory-derived Model 4: Coefficients") %>%
kable_classic(full_width = FALSE, lightable_options = c("striped")) %>%
                    row_spec(row = which(th_4_summ$p.adjusted < 0.05), bold = TRUE)

```

## Omnibus ANOVA test

```{r}
kbl(BIC_summ_anv, caption = "Stepwise BIC Model: ANOVA test") %>%
kable_classic(full_width = FALSE, lightable_options = c("striped")) %>%
                    row_spec(row = which(BIC_summ_anv$`Pr(>F)` < 0.05), bold = TRUE)

kbl(AIC_summ_anv, caption = "Stepwise AIC Model: ANOVA test") %>%
kable_classic(full_width = FALSE, lightable_options = c("striped")) %>%
                    row_spec(row = which(AIC_summ_anv$`Pr(>F)` < 0.05), bold = TRUE)

kbl(th_1_summ_anv, caption = "Theory-derived Model 1: ANOVA test") %>%
kable_classic(full_width = FALSE, lightable_options = c("striped")) %>%
                    row_spec(row = which(th_1_summ_anv$`Pr(>F)` < 0.05), bold = TRUE)

kbl(th_2_summ_anv, caption = "Theory-derived Model 2: ANOVA test") %>%
kable_classic(full_width = FALSE, lightable_options = c("striped")) %>%
                    row_spec(row = which(th_2_summ_anv$`Pr(>F)` < 0.05), bold = TRUE)

kbl(th_3_summ_anv, caption = "Theory-derived Model 3: ANOVA test") %>%
kable_classic(full_width = FALSE, lightable_options = c("striped")) %>%
                    row_spec(row = which(th_3_summ_anv$`Pr(>F)` < 0.05), bold = TRUE)

kbl(th_4_summ_anv, caption = "Theory-derived Model 4: ANOVA test") %>%
kable_classic(full_width = FALSE, lightable_options = c("striped")) %>%
                    row_spec(row = which(th_4_summ_anv$`Pr(>F)` < 0.05), bold = TRUE)

```


```{r eval=FALSE, include=FALSE}
# ggplot(complete(imputed_df), aes(q02_age, PHQ8, color = q11_education)) + geom_point(position = "jitter", alpha = 0.2) + geom_smooth()
# 
# # + geom_parallel_slopes(se = TRUE)
```
